+++ read.go
@@ -322,7 +322,7 @@ func (fdb *FluxDB) ReadSingletEntryAt(
 	if len(key) > 0 && len(value) > 0 {
 		entry, err = NewSingletEntry(singlet, key, value)
 		if err != nil {
-			return nil, fmt.Errorf("failed to create single tablet row %q: %w", Key(key), err)
+			return nil, fmt.Errorf("failed to create singlet entry %q: %w", Key(key), err)
 		}
 	}
 
@@ -343,6 +343,55 @@ func (fdb *FluxDB) ReadSingletEntryAt(
 
 	zlog.Debug("finished reading singlet entry", zap.Bool("entry_exist", entry != nil))
 	return entry, nil
+}
+
+// ReadSingletEntries query the storage engine returning all entries for a precise singlet
+//
+// Returns `<Entry>, nil` when an entry has been found, `nil, nil` when no entry was found
+// and finally, `nil, <error>` if an error was encountered while fetching the singlet entry.
+func (fdb *FluxDB) ReadSingletEntries(
+	ctx context.Context,
+	singlet Singlet,
+	speculativeWrites []*WriteRequest,
+) ([]SingletEntry, error) {
+	ctx, span := dtracing.StartSpan(ctx, "read singlet entries", "singlet", singlet)
+	defer span.End()
+
+	// We are using inverted block num, so we are scanning from highest block num to lowest block (0)
+	startKey := KeyForSingletAt(singlet, math.MaxUint64)
+	endKey := KeyForSingletAt(singlet, 0)
+
+	zlog := logging.Logger(ctx, zlog)
+	zlog.Debug("reading singlet entries from database", zap.Stringer("singlet", singlet), zap.Stringer("start_key", startKey), zap.Stringer("end_key", endKey))
+
+	keys, values, err := fdb.store.FetchSingletEntries(ctx, startKey, endKey)
+	if err != nil {
+		return nil, fmt.Errorf("db fetch single entry: %w", err)
+	}
+
+	entries := make([]SingletEntry, 0, len(keys)+8)
+	for i, key := range keys {
+		entry, err := NewSingletEntry(singlet, key, values[i])
+		if err != nil {
+			return nil, fmt.Errorf("failed to create singlet entry %q: %w", Key(key), err)
+		}
+
+		entries = append(entries, entry)
+	}
+
+	zlog.Debug("reading singlet entry from speculative writes", zap.Int("db_entry_count", len(entries)), zap.Int("speculative_write_count", len(speculativeWrites)))
+	for _, writeRequest := range speculativeWrites {
+		for _, speculativeEntry := range writeRequest.SingletEntries {
+			if !SingletEqual(singlet, speculativeEntry.Singlet()) {
+				continue
+			}
+
+			entries = append(entries, speculativeEntry)
+		}
+	}
+
+	zlog.Debug("finished reading singlet entry", zap.Int("entry_count", len(entries)))
+	return entries, nil
 }
 
 func (fdb *FluxDB) HasSeenAnyRowForTablet(ctx context.Context, tablet Tablet) (exists bool, err error) {